Crate libra_global_constants

The purpose of this crate is to offer a single source of truth for the definitions of shared constants within the Libra codebase. This is useful because many different components within Libra often require access to global constant definitions (e.g., Libra Safety Rules, the Key Manager, and Secure Storage). To avoid duplicating these definitions across crates (and better allow these constants to be updated in a single location), we define them here.

Constants

ASSOCIATION_KEY
Definitions of global cryptographic keys (e.g., as held in secure storage)

CONSENSUS_KEY

EPOCH
Definitions of global data items (e.g., as held in secure storage)

FULLNODE_NETWORK_KEY

LAST_VOTED_ROUND

OPERATOR_KEY

OWNER_KEY

PREFERRED_ROUND

VALIDATOR_NETWORK_KEY

WAYPOINT